+++ apps/backend/src/post/post-query.service.spec.ts
@@ -3,7 +3,8 @@ import { getRepositoryToken } from "@nestjs/typeorm";
 import { DataSource, Repository } from "typeorm";
 import PGMem from "pg-mem";
 import { PostQueryService } from "./post-query.service";
-import { InvalidCursorError, PostNotFoundError } from "./post.errors";
+import { PostNotFoundError } from "./post.errors";
+import { InvalidCursorError } from "src/common/errors/pagination.errors";
 import { Post, PostCategory } from "./entities/post.entity";
 import { PostLike } from "./entities/post-like.entity";
 import { Attachment } from "../attachment/entities/attachment.entity";
